+++ lib/ReactViews/SidePanel/Branding.tsx
@@ -3,25 +3,24 @@ import React from "react";
 import defined from "terriajs-cesium/Source/Core/defined";
 import Terria from "../../Models/Terria";
 import parseCustomHtmlToReact from "../Custom/parseCustomHtmlToReact";
+import ViewState from "../../ReactViewModels/ViewState";
 
 export default (props: {
   terria: Terria;
+  viewState: ViewState;
   version?: string;
-  displayOne?: number; // pass in a number here to only show one item from brandBarElements
-  onClick?: () => void;
 }) => {
-  let brandingHtmlElements = props.terria.configParameters.brandBarElements;
-  if (!defined(brandingHtmlElements)) {
-    brandingHtmlElements = [
+  // Use brandBarElements or brandBarSmallElements depending if using SmallScreenInterface (and default to brandBarElements)
+  let brandingHtmlElements = (props.viewState.useSmallScreenInterface
+    ? props.terria.configParameters.brandBarSmallElements
+    : props.terria.configParameters.brandBarElements) ??
+    props.terria.configParameters.brandBarElements ?? [
       '<a target="_blank" href="http://terria.io"><img src="images/terria_logo.png" height="52" title="Version: {{ version }}" /></a>'
     ];
-  }
-
-  if (!brandingHtmlElements) return null;
 
   const version = props.version ?? "Unknown";
 
-  const displayOne = props.displayOne;
+  const displayOne = props.terria.configParameters.displayOneBrand;
   const displayContent =
     // If the index exists, use that
     (displayOne && brandingHtmlElements[displayOne]) ||
@@ -63,19 +62,10 @@ export default (props: {
           ${(p: any) => p.theme.logoPaddingVertical};
 
         @media (max-width: ${(p: any) => p.theme.sm}px) {
-          height: ${(p: any) => p.theme.inputHeight};
-          padding: 0 ${(p: any) => p.theme.spacing}px;
+          height: ${(p: any) => p.theme.logoSmallHeight};
 
-          // Remove all font-size so text scales to smaller screen
-          * {
-            font-size: unset !important;
-            padding: 0 !important;
-          }
-
-          // Only show first brandingHtmlElement on small screen
-          * :not(:first-child) {
-            display: none;
-          }
+          padding: ${(p: any) => p.theme.logoSmallPaddingHorizontal}
+            ${(p: any) => p.theme.logoSmallPaddingVertical};
 
           // Remove a "display: flex" on small screen
           a {