+++ lib/Sass/exports/_variables-export.scss.d.ts
@@ -23,6 +23,9 @@ interface CssExports {
   'logoHeight': string;
   'logoPaddingHorizontal': string;
   'logoPaddingVertical': string;
+  'logoSmallHeight': string;
+  'logoSmallPaddingHorizontal': string;
+  'logoSmallPaddingVertical': string;
   'mapButtonColor': string;
   'mapButtonTop': string;
   'mapNavigationTop': string;