+++ CHANGES.md
@@ -5,7 +5,8 @@ Change Log
 
 #### next release (8.0.0-alpha.77)
 
-* Convert `Branding.jsx` to `.tsx` + fix small screen branding
+* Convert `Branding.jsx` to `.tsx`
+* Added `configParams.brandBarSmallElements` to set Branding elements for small screen (also added theme props)
 * Add `font` variables and `fontImports` to theme - this can be used to import CSS fonts.
 * Convert `lib/Styled` `.jsx` files to `.tsx` (including Box, Icon, Text). The most significant changes to these interfaces are:
   * `Box` no longer accepts `<Box positionAbsolute/>` and this should now be passed as `<Box position="absolute"/>`.